# src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: API 문서
  description: RestDocs With Swagger
  version: 0.0.1
servers:
- url: http://devterview.com:8080
- url: http://localhost:8080
- url: http://dev-api.devterview.com:8080
- url: http://dev-api.devterview.com
...
paths:
...
  /post:
    post:
      tags:
      - Post API
      summary: 게시글 API
      description: 게시글 API
      operationId: 게시글 작성 테스트(
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/post-881832127'
            examples:
              게시글 작성 테스트(면접 공유 INTERVIEW) - 성공:
                value: |-
                  {
                    "jobKoreanName" : "백엔드",
                    "content" : "This is a test post content.",
                    "s3ImageUrl" : "https://s3.amazonaws.com/test-image.jpg",
                    "interviewId" : 1,
                    "overallEvaluationId" : 2,
                    "postType" : "INTERVIEW"
                  }
              게시글 작성 테스트(일반 POST) - 성공:
                value: |-
                  {
                    "jobKoreanName" : "백엔드",
                    "content" : "This is a test post content.",
                    "s3ImageUrl" : "https://s3.amazonaws.com/test-image.jpg",
                    "interviewId" : 1,
                    "overallEvaluationId" : 2,
                    "postType" : "POST"
                  }
              게시글 작성 테스트(평가 공유 EVALUATION) - 성공:
                value: |-
                  {
                    "jobKoreanName" : "백엔드",
                    "content" : "This is a test post content.",
                    "s3ImageUrl" : "https://s3.amazonaws.com/test-image.jpg",
                    "interviewId" : 1,
                    "overallEvaluationId" : 2,
                    "postType" : "EVALUATION"
                  }
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/post-1902421517'
              examples:
                게시글 작성 테스트(면접 공유 INTERVIEW) - 성공:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "postId" : 100,
                        "jobName" : "BACK_END",
                        "jobNameKorean" : "백엔드",
                        "content" : "This is a test post content.",
                        "s3ImageUrl" : "https://s3.amazonaws.com/test-image.jpg",
                        "interview" : {
                          "interviewId" : 1,
                          "interviewTitle" : "Mock Interview Title",
                          "interviewStatus" : "COMPLETED",
                          "interviewType" : "TECHNICAL",
                          "interviewMethod" : "VIDEO",
                          "interviewMode" : "REAL",
                          "questionCount" : 10,
                          "job" : {
                            "jobId" : 1,
                            "jobName" : "BACK_END",
                            "jobNameKorean" : "백엔드",
                            "jobDescription" : "백엔드 직무입니다."
                          },
                          "files" : [ {
                            "fileId" : 1,
                            "type" : "COVER_LETTER",
                            "fileName" : "cover_letter.pdf",
                            "s3FileUrl" : "https://s3.amazonaws.com/resume.pdf"
                          } ]
                        },
                        "evaluation" : null,
                        "postType" : "INTERVIEW",
                        "generatedAt" : "2024-12-01T01:05:37.058773"
                      }
                    }
                게시글 작성 테스트(일반 POST) - 성공:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "postId" : 100,
                        "jobName" : "BACK_END",
                        "jobNameKorean" : "백엔드",
                        "content" : "This is a test post content.",
                        "s3ImageUrl" : "https://s3.amazonaws.com/test-image.jpg",
                        "interview" : null,
                        "evaluation" : null,
                        "postType" : "POST",
                        "generatedAt" : "2024-12-01T01:05:37.031773"
                      }
                    }
                게시글 작성 테스트(평가 공유 EVALUATION) - 성공:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "postId" : 100,
                        "jobName" : "BACK_END",
                        "jobNameKorean" : "백엔드",
                        "content" : "This is a test post content.",
                        "s3ImageUrl" : "https://s3.amazonaws.com/test-image.jpg",
                        "interview" : null,
                        "evaluation" : {
                          "interview" : {
                            "interviewId" : 1,
                            "interviewTitle" : "Mock Interview Title",
                            "interviewStatus" : "COMPLETED",
                            "interviewType" : "TECHNICAL",
                            "interviewMethod" : "VIDEO",
                            "interviewMode" : "REAL",
                            "questionCount" : 10,
                            "job" : {
                              "jobId" : 1,
                              "jobName" : "BACK_END",
                              "jobNameKorean" : "백엔드",
                              "jobDescription" : "백엔드 직무입니다."
                            },
                            "files" : [ {
                              "fileId" : 1,
                              "type" : "COVER_LETTER",
                              "fileName" : "cover_letter.pdf",
                              "s3FileUrl" : "https://s3.amazonaws.com/resume.pdf"
                            } ]
                          },
                          "evaluationCriteria" : [ {
                            "evaluationCriteriaId" : 1,
                            "evaluationCriteria" : "Communication Skills",
                            "feedbackText" : "Clear and concise.",
                            "score" : 9
                          }, {
                            "evaluationCriteriaId" : 2,
                            "evaluationCriteria" : "Problem-Solving",
                            "feedbackText" : "Good examples provided.",
                            "score" : 8
                          } ],
                          "answerEvaluations" : [ {
                            "answerEvaluationId" : 1,
                            "questionText" : "What is your greatest strength?",
                            "answerText" : "I am highly adaptable.",
                            "answerFeedbackStrength" : "Good emphasis on adaptability.",
                            "answerFeedbackImprovement" : "Could elaborate more.",
                            "answerFeedbackSuggestion" : "Provide an example of adaptability.",
                            "answerEvaluationScores" : [ {
                              "answerEvaluationScoreId" : 1,
                              "answerEvaluationScoreName" : "Clarity",
                              "score" : 8,
                              "rationale" : "Clear explanation",
                              "answerEvaluationType" : "TEXT"
                            }, {
                              "answerEvaluationScoreId" : 2,
                              "answerEvaluationScoreName" : "Conciseness",
                              "score" : 7,
                              "rationale" : "Could be more concise",
                              "answerEvaluationType" : "TEXT"
                            } ]
                          } ]
                        },
                        "postType" : "EVALUATION",
                        "generatedAt" : "2024-12-01T01:05:36.997963"
                      }
                    }
...
  /subscription:
    post:
      tags:
      - Subscription API
      summary: 구독 생성 API
      description: 구독 생성 API
      operationId: 구독 생성 테스트 - 성공
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/subscription-2024782262'
            examples:
              구독 생성 테스트 - 성공:
                value: |-
                  {
                    "jobId" : 2
                  }
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/subscription129725505'
              examples:
                구독 생성 테스트 - 성공:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "subscriptionId" : 1,
                        "userId" : 1,
                        "jobId" : 2,
                        "subscribedAt" : "2024-12-01T01:05:39.419952"
                      }
                    }
...
components:
  schemas:
...
    subscription-2024782262:
      type: object
      properties:
        jobId:
          type: number
          description: 직무 ID
...
    post-1902421517:
      type: object
      properties:
        code:
          type: number
          description: 응답 코드
        data:
          type: object
          properties:
            evaluation:
              type: object
              properties:
                answerEvaluations:
                  type: array
                  items:
                    type: object
                    properties:
                      answerEvaluationId:
                        type: number
                        description: 답변 평가 식별자
                      answerFeedbackImprovement:
                        type: string
                        description: 답변 개선 피드백
                      answerFeedbackSuggestion:
                        type: string
                        description: 답변 제안 피드백
                      answerText:
                        type: string
                        description: 답변 텍스트
                      answerEvaluationScores:
                        type: array
                        items:
                          type: object
                          properties:
                            answerEvaluationType:
                              type: string
                              description: "답변 평가 유형(TEXT, VOICE, VIDEO"
                            score:
                              type: number
                              description: 답변 평가 점수
                            answerEvaluationScoreId:
                              type: number
                              description: 답변 평가 점수 식별자
                            answerEvaluationScoreName:
                              type: string
                              description: 답변 평가 점수 이름
                            rationale:
                              type: string
                              description: 답변 평가 점수 근거
                      questionText:
                        type: string
                        description: 질문 텍스트
                      answerFeedbackStrength:
                        type: string
                        description: 답변 강점 피드백
                evaluationCriteria:
                  type: array
                  items:
                    type: object
                    properties:
                      feedbackText:
                        type: string
                        description: 평가 피드백 텍스트
                      score:
                        type: number
                        description: 평가 점수
                      evaluationCriteria:
                        type: string
                        description: 평가 기준
                      evaluationCriteriaId:
                        type: number
                        description: 평가 기준 식별자
                interview:
                  type: object
                  properties:
                    questionCount:
                      type: number
                      description: 질문 개수
                    interviewStatus:
                      type: string
                      description: 면접 상태
                    interviewId:
                      type: number
                      description: 면접 식별자
                    interviewType:
                      type: string
                      description: 면접 유형
                    interviewTitle:
                      type: string
                      description: 면접 제목
                    interviewMethod:
                      type: string
                      description: 면접 방식
                    files:
                      type: array
                      items:
                        type: object
                        properties:
                          fileName:
                            type: string
                            description: 파일 이름
                          s3FileUrl:
                            type: string
                            description: S3 파일 URL
                          type:
                            type: string
                            description: "파일 유형 (예: COVER_LETTER)"
                          fileId:
                            type: number
                            description: 파일 식별자
                    job:
                      type: object
                      properties:
                        jobName:
                          type: string
                          description: 직무 이름
                        jobId:
                          type: number
                          description: 직무 식별자
                        jobNameKorean:
                          type: string
                          description: 직무 이름 (한글)
                        jobDescription:
                          type: string
                          description: 직무 설명
                    interviewMode:
                      type: string
                      description: 면접 모드
              description: 일반 게시글 작성에서는 평가 정보 없음
            jobName:
              type: string
              description: 직무 이름
            postType:
              type: string
              description: "게시글 유형 (EVALUATION, INTERVIEW, POST)"
            generatedAt:
              type: string
              description: 작성 일시
            jobNameKorean:
              type: string
              description: 직무 이름 (한글)
            s3ImageUrl:
              type: string
              description: S3 이미지 URL
            postId:
              type: number
              description: 게시글 식별자
            interview:
              type: object
              properties:
                questionCount:
                  type: number
                  description: 질문 개수
                interviewStatus:
                  type: string
                  description: 면접 상태
                interviewId:
                  type: number
                  description: 면접 식별자
                interviewType:
                  type: string
                  description: 면접 유형
                interviewTitle:
                  type: string
                  description: 면접 제목
                interviewMethod:
                  type: string
                  description: 면접 방식
                files:
                  type: array
                  items:
                    type: object
                    properties:
                      fileName:
                        type: string
                        description: 파일 이름
                      s3FileUrl:
                        type: string
                        description: S3 파일 URL
                      type:
                        type: string
                        description: "파일 유형 (예: COVER_LETTER)"
                      fileId:
                        type: number
                        description: 파일 식별자
                job:
                  type: object
                  properties:
                    jobName:
                      type: string
                      description: 직무 이름
                    jobId:
                      type: number
                      description: 직무 식별자
                    jobNameKorean:
                      type: string
                      description: 직무 이름 (한글)
                    jobDescription:
                      type: string
                      description: 직무 설명
                interviewMode:
                  type: string
                  description: 면접 모드
              description: 평가 공유 게시글 작성에서는 인터뷰 정보 없음
            content:
              type: string
              description: 게시글 내용
        message:
          type: string
          description: 응답 메시지
...
    post-881832127:
      type: object
      properties:
        overallEvaluationId:
          type: number
          description: 종합 평가 식별자
        interviewId:
          type: number
          description: 면접 식별자
        jobKoreanName:
          type: string
          description: 직무 이름 (한글)
        postType:
          type: string
          description: "게시글 유형 (EVALUATION, INTERVIEW, POST)"
        s3ImageUrl:
          type: string
          description: S3 이미지 URL
        content:
          type: string
          description: 게시글 내용
...
    subscription129725505:
      type: object
      properties:
        code:
          type: number
          description: 응답 코드
        data:
          type: object
          properties:
            jobId:
              type: number
              description: 직무 ID
            subscribedAt:
              type: string
              description: 구독 생성 시각
            subscriptionId:
              type: number
              description: 구독 ID
            userId:
              type: number
              description: 사용자 ID
        message:
          type: string
          description: 응답 메시지
...
  securitySchemes:
    APIKey:
      type: apiKey
      name: access_token
      in: cookie
security:
  - APIKey: []